Add share link feature for sent conversations (#21)
When a conversation is sent, users can now share a direct link with their
parent instead of asking them to visit the site and type a code manually.

Changes:
- Sent confirmation toast now shows a full shareable URL
  (app.ourhearth.co/parent?code=XXXXXX) with a copy-to-clipboard button
  and helper text. The manual code fallback is kept below it.
- Conversations page: sent/in-progress cards now include a copy link button
  alongside the access code badge.
- /parent page: split into a server component that looks up the conversation
  from the ?code= param before rendering, and a client component for
  interactions. When arriving via a shared link, the welcome screen renders
  immediately with no flash or client-side round-trip. Manual code entry
  still works as before.
- CopyLinkButton: reusable client component with clipboard icon and transient
  'Copied!' confirmation state.
